refactor AST types into a separate file
