+++ packages/squiggle-lang/src/ast/parse.ts
@@ -2,45 +2,21 @@ import { ICompileError } from "../errors/IError.js";
 import * as Result from "../utility/result.js";
 import { result } from "../utility/result.js";
 import { SExpr, SExprPrintOptions, sExprToString } from "../utility/sExpr.js";
-import { type ASTCommentNode, type ASTNode } from "./peggyHelpers.js";
 import {
   parse as peggyParse,
   SyntaxError as PeggySyntaxError,
 } from "./peggyParser.js";
-
-export { type ASTNode } from "./peggyHelpers.js";
-
-// Types copy-pasted from Peggy, but converted from interface to type.
-// We need a type because interfaces don't match JsonValue type that we use for serialization.
-
-/** Provides information pointing to a location within a source. */
-export type Location = {
-  /** Line in the parsed source (1-based). */
-  line: number;
-  /** Column in the parsed source (1-based). */
-  column: number;
-  /** Offset in the parsed source (0-based). */
-  offset: number;
-};
-
-/** The `start` and `end` position's of an object within the source. */
-export type LocationRange = {
-  /** Unlike in Peggy, this must be a string. */
-  source: string;
-  /** Position at the beginning of the expression. */
-  start: Location;
-  /** Position after the end of the expression. */
-  end: Location;
-};
+import {
+  AST,
+  type ASTCommentNode,
+  type ASTNode,
+  LocationRange,
+} from "./types.js";
 
 export type ParseError = {
   type: "SyntaxError";
   location: LocationRange;
   message: string;
-};
-
-export type AST = Extract<ASTNode, { type: "Program" }> & {
-  comments: ASTCommentNode[];
 };
 
 type ParseResult = result<AST, ICompileError>;
