+++ packages/squiggle-lang/src/ast/utils.ts
@@ -1,5 +1,4 @@
-import { LocationRange } from "./parse.js";
-import { ASTNode } from "./peggyHelpers.js";
+import { ASTNode, LocationRange } from "./types.js";
 
 export function locationContains(location: LocationRange, offset: number) {
   return location.start.offset <= offset && location.end.offset >= offset;
