+++ packages/squiggle-lang/src/value/valueTags.ts
@@ -1,4 +1,4 @@
-import { LocationRange } from "../ast/parse.js";
+import { LocationRange } from "../ast/types.js";
 import {
   SquiggleDeserializationVisitor,
   SquiggleSerializationVisitor,
